refactor(proxy): give the host-path refusal its own module
Review: the narration I added carried what names, a module boundary, and test
names should carry, and `daemon-proxy.ts` had held no such comments before.

`proxy-install-source-admission.ts` now owns the rule, so the boundary states
it instead of a comment; `isBackedByUploadedArtifact` names what "unbacked"
means in `carriesUnbackedHostPathInstallSource`. The proxy owner returns to 443
lines from 495.

Refs #2097
